Make util aware of the backend

+++ src/project_util/project/project.py
@@ -10,7 +10,7 @@ from loguru import logger
 from PIL import Image
 from vidutil.encoder import VideoEncoder
 
-from project_util.constants import FILE_SYSTEM
+from project_util.constants import FILE_SYSTEM, S3
 from project_util.services.s3 import S3Client
 
 
@@ -19,12 +19,12 @@ TProject = TypeVar("TProject", bound="Project")
 
 class Project:
     def __init__(self, name: str, parent_dir: Path, backend: str = FILE_SYSTEM):
+        self._backend = backend
         self._parent_dir = parent_dir
         self._name = name
         self._project_dir: Path = self._make_project_dir(self._name)
         self.folders: Dict[str, Project] = {}
         self._s3_client = None
-        self._backend = backend
 
     @property
     def s3_client(self):
@@ -37,8 +37,13 @@ class Project:
         # todo: implement
 
     def _make_project_dir(self, name: str) -> Path:
-        path = Path(join(self._parent_dir, name)).absolute()
-        makedirs(path, exist_ok=True)
+        if self._backend == FILE_SYSTEM:
+            path = Path(join(self._parent_dir, name)).absolute()
+            makedirs(path, exist_ok=True)
+        elif self._backend == S3:
+            path = Path(join(self._parent_dir, name))
+        else:
+            raise ValueError(f"Unsupported backend: {self._backend}")
         return path
 
     @property
@@ -74,11 +79,18 @@ class Project:
         logger.debug(f"Loading image paths: {paths}")
         return VideoEncoder.load_images(paths)
 
+    def _require_backend(self, backend):
+        if self._backend != backend:
+            raise RuntimeError(
+                f"operation is only supported on backend {backend}"
+            )
+
     def save_image(
         self,
         data: np.ndarray,
         file_name: Path,
     ) -> str:
+        self._require_backend(FILE_SYSTEM)
         path = os.path.join(self.path, file_name)
         os.makedirs(os.path.dirname(path), exist_ok=True)
         im = Image.fromarray(np.uint8(data))
@@ -111,6 +123,7 @@ class Project:
         :param name: file name to export
         :return:
         """
+        self._require_backend(FILE_SYSTEM)
         if target_project:
             target_path = target_project.path
         else:
@@ -127,11 +140,14 @@ class Project:
         )
 
     def add_folder(self, name: str) -> TProject:
-        folder = Project(name, parent_dir=self._project_dir)
+        folder = Project(
+            name, parent_dir=self._project_dir, backend=self._backend
+        )
         self.folders[name] = folder
         return folder
 
     def remove_folder(self, name: str) -> None:
         # todo: use abspath
-        os.removedirs(name)
+        if self._backend == FILE_SYSTEM:
+            os.removedirs(name)
         del self.folders[name]
